fix: 삭제 시 화면 표시
삭제 시 바로 업데이트 안되는 문제 해결 (selectedDate.tsx 부분)

+++ src/components/SelectedDate.tsx
@@ -149,7 +149,10 @@ function SelectedDate({ selectedId }: ISelectedDate) {
     setForm((prev) => !prev);
   };
   const deletePost = async () => {
+    const ok = window.confirm("정말로 삭제하시겠습니까?");
+    if (ok && userId) {
       await deleteDoc(doc(db, selectedId, userId));
+      setSelectedData(null);
       const docRef = doc(db, userId, selectedId.slice(0, 6));
       const docSnap = await getDoc(docRef);
       if (docSnap.exists()) {
